Create views: Login, Splash, OnBoarding(50%)
